feat: enhanced merge sort
Removed a for loop that was added O(n) to the time of the algorithm, and
utilize the build in slice of slice method to separate the two halves of
the original slice.

Exported the Merge function as well, so that if you want to merge 2
sorted slices you already have, that can be done.